requirements.txt
fire>=0.1.3
regex==2017.4.5
gpt2-pytorch_model.bin